LayoutTests/imported/w3c/web-platform-tests/websockets/back-forward-cache-with-open-websocket-connection.window-expected.txt


FAIL Testing BFCache support for page with open WebSocket connection. assert_equals: document unexpectedly BFCached expected (undefined) undefined but got (boolean) true
